backend_apis/requirements.txt
# Copyright 2023 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

google-cloud-aiplatform
google-cloud-aiplatform[preview]
fastapi
uvicorn
Pillow
newspaper3k
pandas
db-dtypes
google-cloud-datacatalog
google-cloud-discoveryengine
google-api-python-client
python-multipart
firebase-admin
pyrebase4
google-cloud-secret-manager
google-cloud-translate
google-cloud-storage
google-cloud-texttospeech
google-auth
lxml[html_clean]